Moved to Visual Studio 20015 (Updated premake to support it) Added first pass integration of imGUI UI library (*nice*) Added support from scissoring per drawcall Added quickly some sorting of drawcalls Fixed some memory leak in resource creation Resource with refcount 0, freed at the end of the current frame Converted Hash & Flag classes to use template rather than Preprocessor Macro hacks First step to move awat from the proxy resource class

+++ Engine/ZenCore/Resource/Graphic/zcResGfxTexture_DX11.cpp
@@ -35,7 +35,7 @@ bool GfxTexture2dProxy_DX11::Initialize(class GfxTexture2d& _Owner)
 	bool bIsDepth							= zcMgr::GfxRender.IsDepth(rResData->meFormat);
 	bufferDesc.Width						= rResData->maMipData[0].mvDim.x;
 	bufferDesc.Height						= rResData->maMipData[0].mvDim.y;
-	bufferDesc.MipLevels					= rResData->maMipData.Count();
+	bufferDesc.MipLevels					= static_cast<UINT>(rResData->maMipData.Count());
 	bufferDesc.ArraySize					= 1;
 	bufferDesc.Format						= zcMgr::GfxRender.ZenFormatToNative(rResData->meFormat);
 	bufferDesc.SampleDesc.Count				= 1;
